chore: update naming, metadata, and readme
