+++ poetry.toml
@@ -1,2 +1,3 @@
 [virtualenvs]
 in-project = true
+path = ".venv"
