+++ pyproject.toml
@@ -1,11 +1,13 @@
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
 [tool.poetry]
-name = "idex-sdk-python"
-version = "0.1.0"
-description = ""
-authors = ["Your Name <you@example.com>"]
-packages = [
-    { include = "idex_sdk_python" },
-]
+name = "idex-sdk"
+version = "1.0.0a1"
+description = "IDEX v3 SDK for Python"
+authors = ["IDEX <support@idex.io>"]
+packages = [{ include = "idex_sdk" }]
 
 [tool.poetry.dependencies]
 python = ">3.7.1,<3.10"
@@ -21,11 +23,7 @@ black = "^22.3.0"
 types-requests = "^2.27.30"
 python-dotenv = "^0.20.0"
 debugpy = "^1.6.2"
-pydoc-markdown = {version = "^4.6.3", extras = ["novella"]}
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+pydoc-markdown = { version = "^4.6.3", extras = ["novella"] }
 
 [tool.mypy]
 warn_unused_configs = true
@@ -37,6 +35,3 @@ include = '\.pyi?$'
 
 [tool.isort]
 profile = "black"
-
-[virtualenvs]
-path = ".venv"
